Adding this project's main code
Here is the addition of the code used in this project. I added the main_all_analysis.py for those scientists who wish to do a complete analysis of their data (you must have the tables and images ready). But I also added the 3 main functions of the code separately for those who wish to go over only one of the sections (data analysis, image processing, or statistics). The goal is to facilitate the use of this tool for those unfamiliar with coding.
Hope this code is clear to all wishing to use it.